+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
-1.0.0-beta.5-SNAPSHOT (2021-11-22)
+1.0.0-beta.5 (2021-12-07)
 ---------------------------
 
 **Added**